% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crypto_quotes.R
\name{get_crypto_quotes}
\alias{get_crypto_quotes}
\title{Scrap crypto quotes data}
\usage{
get_crypto_quotes(
  crypto_id = 1,
  currency = "USD",
  interval = "1h",
  time_start = Sys.Date() - 5,
  time_end = Sys.Date()
)
}
\arguments{
\item{crypto_id}{Id of crypto asset}

\item{currency}{In which currency to show value of the crypto asset}

\item{interval}{Time interval; can be '1h' or '1d'}

\item{time_start}{date objecct, start of series}

\item{time_end}{date objecct, end of series}
}
\value{
data.table of quotes and metadata
}
\description{
Scrap crypto quotes data
}
